Initial commit of base after renaming
This commit copies the code from the old "conventions" module. I am
copying the code because I do not want to delete the repository just
yet. It is used in the paper's examples.

My new naming standard is to have the following:

 * <modulename>-loc - all morloc definitions, no implementations
 * <modulename>-<lang> - language-specific definitions
 * <modulename> - a wrapper module that imports all lang modules

I need to update the parser and all the highlighters to allow hyphens in
module names.

```morloc
module base-py (*)

import base-loc